Widen device_commands quota_reserved_{memory,storage}_bytes to bigint
Both store a byte count (a parsed memory_limit/storage_limit), like
devices.allocated_memory_bytes/used_memory_bytes/allocated_storage_bytes/used_storage_bytes
already do as bigint - these two were left as a 32-bit integer (max ~2.1GB) in the migration
that added this table. Real, production-blocking bug: resuming any container with a memory or
storage limit at or above ~2GB failed outright with psycopg2.errors.NumericValueOutOfRange,
since the INSERT into device_commands couldn't hold the value at all. Applied directly to prod
via ALTER TABLE to unblock immediately; this commit is the matching model + migration so a
fresh database (and any future migration run) gets it too.

+++ browseterm_db/models/device_commands.py
@@ -13,7 +13,7 @@ from datetime import datetime, timezone
 from typing import Dict, Any
 
 # sqlalchemy
-from sqlalchemy import Column, String, Integer, DateTime, Index, ForeignKey, Enum, and_
+from sqlalchemy import Column, String, Integer, BigInteger, DateTime, Index, ForeignKey, Enum, and_
 from sqlalchemy.dialects.postgresql import UUID, JSON
 from sqlalchemy.orm import relationship
 
@@ -110,8 +110,12 @@ class DeviceCommand(Base):
     # Quota accounting - exactly-once release guard (see DeviceCommandOps.release_quota_for_command).
     # Nullable: only CREATE/RESUME commands ever reserve quota in the first place.
     quota_reserved_cpu = Column(Integer, nullable=True)
-    quota_reserved_memory_bytes = Column(Integer, nullable=True)
-    quota_reserved_storage_bytes = Column(Integer, nullable=True)
+    # Byte counts, not core counts - matching devices.allocated_memory_bytes/used_memory_bytes/
+    # allocated_storage_bytes/used_storage_bytes' own BigInteger type (see the migration that
+    # widened these from a plain Integer for why: a 32-bit column can't hold a routine multi-GB
+    # limit at all).
+    quota_reserved_memory_bytes = Column(BigInteger, nullable=True)
+    quota_reserved_storage_bytes = Column(BigInteger, nullable=True)
     quota_released_at = Column(DateTime, nullable=True)
 
     created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc), nullable=False)
